backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
redis==5.0.1
websockets==12.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2
